refactor: moved prompt.py to a module
