+++ src/ragas/prompt/base.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import inspect
 import typing as t
 from abc import ABC, abstractmethod
 
@@ -150,25 +149,3 @@ class StringPrompt(BasePrompt):
             callbacks=callbacks,
         )
         return llm_result.generations[0][0].text
-
-
-class PromptMixin:
-    def get_prompts(self) -> t.Dict[str, PydanticPrompt]:
-        prompts = {}
-        for name, value in inspect.getmembers(self):
-            if isinstance(value, PydanticPrompt):
-                prompts.update({name: value})
-        return prompts
-
-    def set_prompts(self, **prompts):
-        available_prompts = self.get_prompts()
-        for key, value in prompts.items():
-            if key not in available_prompts:
-                raise ValueError(
-                    f"Prompt with name '{key}' does not exist. Use get_prompts() to see available prompts."
-                )
-            if not isinstance(value, PydanticPrompt):
-                raise ValueError(
-                    f"Prompt with name '{key}' must be an instance of 'ragas.prompt.PydanticPrompt'"
-                )
-            setattr(self, key, value)
